colqwen2/docker/requirements.txt
# ColQwen2 dependencies
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
colpali-engine>=0.1.0
pillow>=10.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
qdrant-client>=1.6.0
python-multipart>=0.0.6
# Quantization (optional)
bitsandbytes>=0.41.0
